+++ openstack/rds/v3/configurations/results.go
@@ -38,28 +38,6 @@ type Parameter struct {
 	Description string `json:"description"`
 }
 
-type ApplyConfiguration struct {
-	// Specifies the parameter template ID.
-	ConfigurationID string `json:"configuration_id"`
-	// Specifies the parameter template name.
-	ConfigurationName string `json:"configuration_name"`
-	// Specifies the result of applying the parameter template.
-	ApplyResults []ApplyConfigurationResult `json:"apply_results"`
-	// Specifies whether each parameter template is applied to DB instances successfully.
-	Success bool `json:"success"`
-}
-
-type ApplyConfigurationResult struct {
-	// Indicates the DB instance ID.
-	InstanceID string `json:"instance_id"`
-	// Indicates the DB instance name.
-	InstanceName string `json:"instance_name"`
-	// Indicates whether a reboot is required.
-	RestartRequired bool `json:"restart_required"`
-	// Indicates whether each parameter template is applied to DB instances successfully.
-	Success bool `json:"success"`
-}
-
 func (r CreateResult) ExtractInto(v interface{}) error {
 	return r.Result.ExtractIntoStructPtr(v, "configuration")
 }
@@ -99,32 +77,4 @@ type GetResult struct {
 // method to determine if the request succeeded or failed.
 type DeleteResult struct {
 	golangsdk.ErrResult
-}
-
-// Extract is a function that accepts a result and extracts a list of configurations.
-func (r ListResult) Extract() ([]Configuration, error) {
-	var a struct {
-		Configurations []Configuration `json:"configurations"`
-	}
-	err := r.Result.ExtractInto(&a)
-	return a.Configurations, err
-}
-
-// ListResult represents the result of a list operation. Call its Extract
-// method to interpret it as a list of Configurations.
-type ListResult struct {
-	golangsdk.Result
-}
-
-// Extract is a function that accepts a result and extracts an apply configuration result.
-func (r ApplyResult) Extract() (*ApplyConfiguration, error) {
-	var response ApplyConfiguration
-	err := r.ExtractInto(&response)
-	return &response, err
-}
-
-// ApplyResult represents the result of a apply operation. Call its Extract
-// method to interpret it.
-type ApplyResult struct {
-	golangsdk.Result
 }
